based CRUD

+++ backend/dentist-office/src/main/java/com/project/dentistoffice/repository/RoleRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     //    List<Role> findByName(String name);
-    Optional<Role> findById(Integer id);
+    Optional<Role> findById(Long id);
     Role findByName(String name);
 }
